requirements.txt
Django==1.10.6
djangorestframework=3.6.2
django-debug-toolbar
django-test-plus==1.0.17